Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -10,6 +10,6 @@ int main(){
     }else if(c=='*'){
         printf("%d",a*b);
     }else{
-        printf("%f",(double)a/b);
+        printf("%d",a/b);
     }
 }
